Overhaul of the Qt layout of the GUI
Refashioned much of the layouts, and removed all of the blank
'notes' used as spacers in the layout.  The orientation was
changed to fit much better for a 16:9 aspect ratio, and the
minimium window size was reduced to below 1280x720 pixels.  As this
minimum size should easily fit in any modern 1920x1080 pixel
display, the scroll area functionality was removed to allow for
easier resizing when using larger windows.  The removal of the
QScrollArea may need to be reverted if the lack of scroll bars
causes issues on low resolution displays.  Several tweaks to the
labels were performed as well.